+++ src/main/java/org/mifos/connector/channel/api/definition/GSMATransactionApi.java
@@ -1,7 +1,7 @@
 package org.mifos.connector.channel.api.definition;
 
 import com.fasterxml.jackson.core.*;
-import org.mifos.connector.channel.GSMA_API.*;
+import org.mifos.connector.channel.utils.GsmaP2PResponseDto;
 import org.mifos.connector.common.gsma.dto.*;
 import org.springframework.web.bind.annotation.*;
