+++ src/main/java/org/mifos/connector/channel/api/definition/TransactionApi.java
@@ -1,7 +1,7 @@
 package org.mifos.connector.channel.api.definition;
 
 import com.fasterxml.jackson.core.*;
-import org.mifos.connector.channel.GSMA_API.*;
+import org.mifos.connector.channel.utils.GsmaP2PResponseDto;
 import org.mifos.connector.common.channel.dto.TransactionChannelRequestDTO;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
