+++ src/main/java/org/mifos/connector/channel/api/implementation/CollectionApiController.java
@@ -5,11 +5,14 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.camel.Exchange;
 import org.apache.camel.ProducerTemplate;
 import org.mifos.connector.channel.api.definition.CollectionApi;
-import org.mifos.connector.channel.GSMA_API.GsmaP2PResponseDto;
+import org.mifos.connector.channel.utils.GsmaP2PResponseDto;
 import org.mifos.connector.channel.model.CollectionRequestDTO;
 import org.mifos.connector.channel.utils.Headers;
 import org.mifos.connector.channel.utils.SpringWrapperUtil;
+import org.mifos.connector.common.channel.dto.PhErrorDTO;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
 import java.util.concurrent.*;
 
@@ -23,7 +26,7 @@ public class CollectionApiController implements CollectionApi {
     private ObjectMapper objectMapper;
 
     @Override
-    public GsmaP2PResponseDto collection(String tenant, String correlationId, CollectionRequestDTO requestBody) throws ExecutionException, InterruptedException, JsonProcessingException {
+    public ResponseEntity<Object> collection(String tenant, String correlationId, CollectionRequestDTO requestBody) throws ExecutionException, InterruptedException, JsonProcessingException {
         Headers headers = new Headers.HeaderBuilder()
                 .addHeader("Platform-TenantId", tenant)
                 .addHeader("X-CorrelationID", correlationId)
@@ -33,6 +36,14 @@ public class CollectionApiController implements CollectionApi {
         Exchange ex = producerTemplate.send("direct:post-collection", exchange);
 
         String body = exchange.getIn().getBody(String.class);
-        return objectMapper.readValue(body, GsmaP2PResponseDto.class);    }
+        try {
+            if (body.contains("error")) {
+                return new ResponseEntity<>(objectMapper.readValue(body, PhErrorDTO.class), HttpStatus.BAD_REQUEST);
+            }
+            return new ResponseEntity<>(objectMapper.readValue(body, GsmaP2PResponseDto.class), HttpStatus.OK);
+        } catch (Exception e) {
+            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
 
 }
