+++ src/main/java/org/mifos/connector/channel/utils/GsmaP2PResponseDto.java
@@ -1,4 +1,4 @@
-package org.mifos.connector.channel.GSMA_API;
+package org.mifos.connector.channel.utils;
 
 public class GsmaP2PResponseDto {
     public String transactionId;
